Code based around QUICKSEARCH that allows for the use of multiple tabs independently: - Each search is independent in a tab - Last Multi and Last Solo are independent - It takes care of changes to the saved search list from deleting resources and navigation back to either the list or the next resource in the list - It takes care of opening several options in a new tab - It (should) take account of browsers that do not have javascript’s sessionStorage object - It is optional. Currently, it must be turned on by directly editing configBrowserTagID in the config table – when you (Stéphane) are done with CONFIGURE.php, then I will add the switch there.
I have not done anything yet to the LISTADDTO code (i.e. organizing elements of a list) so expect unusual behaviour if trying that.

It’s complicated by operating in parallel with PHP sessions. If configBrowserTagID is set to 0 or sessionStorage is not supported then PHP sessions are used. This is a lot of extra code and complication. Ultimately, should no major problems be encountered, then I would remove the option to turn it off and would expect everyone to have a browser that supports sessionStorage.

git-svn-id: svn+ssh://svn.code.sf.net/p/wikindx/svn@882 b5da76e6-9d34-46e5-a24a-8866c81a9b49