Hardcode WIKINDX_DB_TABLEPREFIX
It is no longer possible to modify the prefix of the tables in the database. This functionality is only useful for several programs sharing the same database with the same table names or several wikindx installations in the same database. These two practices are to be avoided because they are a good way to lost your data. Each software should be isolated in its own database for privacy, security, bug resistance and ease of maintenance. Firstly, the "wkx_" prefix will be kept for some versions and cannot be changed. Secondly, it will be deleted.

git-svn-id: svn+ssh://svn.code.sf.net/p/wikindx/svn@908 b5da76e6-9d34-46e5-a24a-8866c81a9b49